+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Modified
 - blocks structure for Daggerheart characters
+- languages management for all characters
 
 ## [0.4.5] - 2026-01-30
 ### Added